requirements.txt
nonebot2[fastapi]
nonebot-adapter-onebot
nonebot_plugin_htmlrender==0.6.0
nonebot_plugin_picstatus==2.1.3.post1
nonebot_plugin_apscheduler
tenacity
matplotlib
jieba
requests
openai
pandas
httpx
aiohttp
wordcloud
aiosmtplib
mail
mcrcon
ascii_magic
PicImageSearch
aiorpcx
asteval
yt-dlp
imageio[ffmpeg]
pilmoji[requests]
zhon
orjson
tencentcloud-sdk-python
ffmpeg-python
pdf2image   # also need to install poppler-utils
uvloop
svgwrite
setproctitle
bypy
colour-science
opencv-python
zstandard
aiosqlite
rapidfuzz
sekai_deck_recommend_cpp
pydub
audioop-lts
yappi
brotli
scipy
mistune
psutil<7.0.0

